+++ src/main/java/com/royware/corona/dashboard/services/chart/config/makers/ChartConfigRateOfChangeOfDeaths.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.Map;
 
 import com.royware.corona.dashboard.enums.data.MovingAverageSizes;
-import com.royware.corona.dashboard.interfaces.charts.IChartConfigMaker;
+import com.royware.corona.dashboard.interfaces.chartconfig.IChartConfigMaker;
 import com.royware.corona.dashboard.model.dashboard.DashboardChartConfig;
 
 public class ChartConfigRateOfChangeOfDeaths implements IChartConfigMaker {
