+++ src/main/java/com/royware/corona/dashboard/services/dashboard/DashboardChartServiceImpl.java
@@ -12,8 +12,8 @@ import org.springframework.stereotype.Component;
 import com.royware.corona.dashboard.enums.charts.ChartCsvHeaders;
 import com.royware.corona.dashboard.enums.charts.ChartTypes;
 import com.royware.corona.dashboard.enums.dashstats.DashStatsTypes;
-import com.royware.corona.dashboard.interfaces.charts.IChartConfigStore;
-import com.royware.corona.dashboard.interfaces.charts.IChartListStore;
+import com.royware.corona.dashboard.interfaces.chartconfig.IChartConfigStore;
+import com.royware.corona.dashboard.interfaces.chartlist.IChartListStore;
 import com.royware.corona.dashboard.interfaces.dashboard.DashboardChartService;
 import com.royware.corona.dashboard.interfaces.dashboard.IDashStatsStore;
 import com.royware.corona.dashboard.interfaces.model.CanonicalCaseDeathData;
